tempo metric refactor
